Add missing  param and fix remoting issue.
In commit 69e3a87, documentation and usage of a  parameter was added to FilesystemRemote.ps1, but the actual parameter was not.

Also, inside the remote scriptblock, the code uses Get-Hash, a function private to the module which will not exist ina remote session. If the target machine has the Pscx module installed, it also has a Get-Hash function which could explain why this continued to work, but it should not be relied on. I added code to allow the injection of any local functions into the remote script block. It uses Invoke-Expression, which is not a recommended practice, but it doesn't execute anything that is user entered.